pyproject.toml
[tool.black]
# Максимальная длина строки
line-length = 119
exclude = '''
(\.git)
'''

[tool.isort]
# Максимальная длина строки
line_length = 119

[tool.mypy]
# Запрет на объявление функции без аннотации типов
disallow_untyped_defs = true
# Без этой опции аргументы функции могут иметь тип None в качестве параметра по умолчанию
no_implicit_optional = true
# mypy предупреждает, если в качестве возвращаемого типа было указано Any
warn_return_any = false
exclude = 'venv|tests'
# Проверка корректности типизации в функции без аннотаций
check_untyped_defs = true
# Сообщает, если какой-то кусок кода никогда не выполняется
warn_unreachable = true
# Игнорирует локальный импорт
ignore_missing_imports = true

[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
